fix bug & add favicon
