+++ templates/dashboard/base.html
@@ -10,6 +10,7 @@ scratch. This page gets rid of all links and provides the needed markup only.
   <title>KunLun-MA | {% block title %}index{% endblock %}</title>
     {% load static %}
   <!-- Tell the browser to be responsive to screen width -->
+  <link rel="icon" href="{% static 'favicon.ico' %}" >
   <meta content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no" name="viewport">
   <link rel="stylesheet" href="{% static 'css/bootstrap.min.css' %}">
   <!-- Font Awesome -->
