+++ templates/login.html
@@ -5,7 +5,7 @@
 	<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
 
     {% load static %}
-    <link rel="stylesheet" href="{% static 'css/bulma.css' %}">
+    <link rel="icon" href="{% static 'favicon.ico' %}" ><link rel="stylesheet" href="{% static 'css/bulma.css' %}">
     <link rel="stylesheet" href="{% static 'css/default.css' %}">
 
 	<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
